GP-3324:  Updated the lifecycle to better deal with cli args.
The CLI stuff is proving to be a pain in the ass based on how Bash deals
with arrays.  To avoid problems with that, we're going to cheat and deal
with things a little more "bashy" so that they work.  It just requires
that we introduce more lifecycle than we intended to do originally.